pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pysurf-detector"
version = "0.2.0"
description = "A pure Python implementation of SURF (Speeded-Up Robust Features)"
readme = "README.md"
authors = [
    {name = "Joel Ibaceta", email = "joelibaceta@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["surf", "computer-vision", "feature-detection", "keypoint-detection"]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "matplotlib>=3.3.0",
    "pillow>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/joelibaceta/pySURF"
Documentation = "https://github.com/joelibaceta/pySURF/blob/main/README.md"
Repository = "https://github.com/joelibaceta/pySURF"
"Bug Tracker" = "https://github.com/joelibaceta/pySURF/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pysurf*"]
exclude = ["tests*", "examples*", "env*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | env
  | build
  | dist
)/
'''
